fix(nexus): add bootstrap case

+++ nexus/build.sh
@@ -1,16 +1,26 @@
 . ../meta/functions.sh
 
 CONTAINER_UUID=$(cat /proc/sys/kernel/random/uuid)
-buildah from --name=${CONTAINER_UUID} ${REGISTRY}/openjdk8-jre:$(date +'%Y.%m.%d')-1
+
+if [[ ! -z ${IMAGE_BOOTSTRAP} ]]; then
+    buildah from --name=${CONTAINER_UUID} ${REGISTRY}/bootstrap/openjdk8-jre:$(date +'%Y.%m.%d')-1
+else
+    buildah from --name=${CONTAINER_UUID} ${REGISTRY}/openjdk8-jre:$(date +'%Y.%m.%d')-1
+fi
 CONTAINER_PATH=$(buildah mount ${CONTAINER_UUID})
 
-NEXUS_VERSION="3.32.0-03"
+NEXUS_VERSION="3.33.0-01"
 
 TMPDIR=$(mktemp -d)
 
-pushd ${TMPDIR}
-curl -L http://10.88.0.249:8081/repository/raw-hosted-stuff/nexus/${NEXUS_VERSION}/nexus-${NEXUS_VERSION}-unix.tar.gz|tar xzv
-popd
+if [[ ! -z ${IMAGE_BOOTSTRAP} ]]; then
+    tar xvf files/nexus-${NEXUS_VERSION}-unix.tar.gz -C ${TMPDIR}
+else
+    pushd ${TMPDIR}
+    curl -L https://download.sonatype.com/nexus/3/nexus-${NEXUS_VERSION}-unix.tar.gz|tar xzv
+    #curl -L http://10.88.0.249:8081/repository/raw-hosted-stuff/nexus/${NEXUS_VERSION}/nexus-${NEXUS_VERSION}-unix.tar.gz|tar xzv
+    popd
+fi
 
 mkdir ${CONTAINER_PATH}/usr/lib/sonatype
 pushd ${CONTAINER_PATH}/usr/lib/sonatype
@@ -23,13 +33,19 @@ rm -rf ${TMPDIR}
 
 rsync_rootfs
 
-cp -v files/nexus.vmoptions ${CONTAINER_PATH}/usr/lib/sonatype/nexus/bin/nexus.vmoptions
-cp -v files/keystore.p12 ${CONTAINER_PATH}/usr/lib/sonatype/sonatype-work/nexus3/keystores/keystore.p12
+if [[ -f files/keystore.p12 && -f files/nexus.vmoptions ]]; then
+    cp -v files/nexus.vmoptions ${CONTAINER_PATH}/usr/lib/sonatype/nexus/bin/nexus.vmoptions
+    cp -v files/keystore.p12 ${CONTAINER_PATH}/usr/lib/sonatype/sonatype-work/nexus3/keystores/keystore.p12
+fi
 
 buildah run -t ${CONTAINER_UUID} systemctl enable\
  nexus.service
 
 buildah config --volume /var/lib/sonatype-work ${CONTAINER_UUID}
 
-buildah commit ${CONTAINER_UUID} ${REGISTRY}/nexus:$(date +'%Y.%m.%d')-1
+if [[ ! -z ${IMAGE_BOOTSTRAP} ]]; then
+    buildah commit ${CONTAINER_UUID} ${REGISTRY}/bootstrap/nexus:$(date +'%Y.%m.%d')-1
+else
+    buildah commit ${CONTAINER_UUID} ${REGISTRY}/nexus:$(date +'%Y.%m.%d')-1
+fi
 buildah rm -a
